+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Flip the sign".
 - Tests for "The michelangelo code".
 - Tests for "Gold packing".
+- Tests for "The broken editor".
 
 ## [1.9.0] - 2022-06-01
 ### Added